Move logics, which detect the state update and control `transition`CSS, into the child
